[AMUXCTL] Split code of amuxctl for future introduction of libamuxctl
